+++ meal_credit_webapp/src/index.js
@@ -1,80 +1,12 @@
 import "./css/index.css";
 
-import {
-    Link,
-    Redirect,
-    Route,
-    BrowserRouter as Router,
-    Switch
-} from "react-router-dom";
-
-import APIWrapper from "./API/api_wrapper";
-import CookiesWrapper from "./API/cookies";
-import Home from "./Pages/Home";
-import Login from "./Pages/Login";
-import LogoutButton from "./Pages/Logout"
-import { Provider } from "react-redux";
+import App from "./App";
 import React from "react";
 import ReactDOM from "react-dom";
-import Register from "./Pages/Register";
-import { createStore } from "redux";
-import reducers from "./redux/reducers/index";
 
-document.title = "Meal Credit App";
+ReactDOM.render(<App />, document.getElementById("root"));
 
-var apiWrapper = new APIWrapper();
-var cookieWrapper = new CookiesWrapper();
-var retrievedLoginInfo = cookieWrapper.retrieveCookieIfExists(
-    "user_information"
-);
-const store = createStore(reducers);
-console.log(JSON.parse(retrievedLoginInfo))
-
-if (retrievedLoginInfo !== null) {
-    console.log("NON NULL")
-    store.dispatch({
-        type: "CHANGE_LOGIN_INFO",
-        loginInfo: JSON.parse(retrievedLoginInfo),
-    });
-}
-
-ReactDOM.render(
-    <Provider store={store}>
-        <Router>
-            <Switch>
-                { console.log("fsejkfe") }
-                <Route exact path="/">
-                    { console.log("daw", store.getState().userLoginInfo) }
-                    { store.getState().userLoginInfo ? <Home /> : <Redirect to="/login" /> }
-                    <LogoutButton store={store} cookieWrapper={cookieWrapper} />
-                </Route>
-                <Route path="/login">
-                    {
-                        store.getState().userLoginInfo ? 
-                        <Redirect to="/" /> : 
-                        <Login
-                            cookieWrapper={cookieWrapper}
-                            store={store}
-                            apiWrapper={apiWrapper}
-                        />
-                    }
-                    <LogoutButton store={store} cookieWrapper={cookieWrapper} />
-                    {/* <LogoutButton store={store} cookieWrapper={cookieWrapper} /> */}
-                </Route>
-                <Route path="/register">
-                    {/* <LogoutButton store={store} cookieWrapper={cookieWrapper} /> */}
-                    <Register
-                        cookieWrapper={cookieWrapper}
-                        store={store}
-                        apiWrapper={apiWrapper}
-                    />
-                </Route>
-                <Route path="/home">
-                    { console.log("da321312321w", store.getState().userLoginInfo) }
-                    {/* store.getState().userLoginInfo ? <Home /> : <Redirect to="/login" /> */}
-                </Route>
-            </Switch>
-        </Router>
-    </Provider>,
-    document.getElementById("root")
-);
+// // If you want your app to work offline and load faster, you can change
+// // unregister() to register() below. Note this comes with some pitfalls.
+// // Learn more about service workers: https://bit.ly/CRA-PWA
+// serviceWorker.unregister();
